+++ app/src/main/java/com/codingblocks/cbonlineapp/campaign/CampaignViewModel.kt
@@ -10,7 +10,7 @@ import androidx.paging.LivePagedListBuilder
 import androidx.paging.PagedList
 import com.codingblocks.cbonlineapp.baseclasses.BaseCBViewModel
 import com.codingblocks.cbonlineapp.util.extensions.runIO
-import com.codingblocks.cbonlineapp.util.savedStateValue
+import com.codingblocks.cbonlineapp.util.extensions.savedStateValue
 import com.codingblocks.onlineapi.ResultWrapper
 import com.codingblocks.onlineapi.fetchError
 import com.codingblocks.onlineapi.models.SpinResponse
