+++ app/src/main/java/com/codingblocks/cbonlineapp/course/CourseActivity.kt
@@ -38,6 +38,7 @@ import com.codingblocks.cbonlineapp.util.extensions.getLoadingDialog
 import com.codingblocks.cbonlineapp.util.extensions.getSpannableSring
 import com.codingblocks.cbonlineapp.util.extensions.setRv
 import com.codingblocks.cbonlineapp.util.extensions.setToolbar
+import com.codingblocks.cbonlineapp.util.glide.GlideApp
 import com.codingblocks.cbonlineapp.util.glide.loadImage
 import com.codingblocks.cbonlineapp.util.livedata.observer
 import com.codingblocks.onlineapi.ErrorStatus
@@ -136,7 +137,7 @@ class CourseActivity : BaseCBActivity(), AnkoLogger, AppBarLayout.OnOffsetChange
                 }
             }
         }
-
+        GlideApp.with(this).load(R.drawable.wildcraft).into(goodiesImg)
         viewModel.course.observer(this) { course ->
             endLink = course.slug.toString()
             showTags(course.tags)
