+++ app/src/main/java/com/codingblocks/cbonlineapp/dashboard/explore/DashboardExploreFragment.kt
@@ -86,9 +86,9 @@ class DashboardExploreFragment : BaseCBFragment() {
         vm.fetchRecommendedCourses(0, 4)
         vm.fetchRecommendedCourses(4, 4)
         vm.fetchTracks()
-        campaignView.setOnClickListener {
-            startActivity(CampaignActivity.createCampaignActivityIntent(requireContext()))
-        }
+//        campaignView.setOnClickListener {
+//            startActivity(CampaignActivity.createCampaignActivityIntent(requireContext()))
+//        }
         dashboardPopularRv.setRv(requireContext(), coursePopularListAdapter, orientation = RecyclerView.HORIZONTAL, space = 28f)
         courseSuggestedRv.setRv(requireContext(), courseCardListAdapter, orientation = RecyclerView.HORIZONTAL, space = 28f)
         dashboardTracksRv.setRv(requireContext(), tracksListAdapter, orientation = RecyclerView.HORIZONTAL, space = 28f)
