+++ app/src/main/java/com/codingblocks/cbonlineapp/library/LibraryListAdapter.kt
@@ -40,7 +40,18 @@ class LibraryListAdapter(val type: LibraryTypes) : ListAdapter<BaseModel, Recycl
         setHasStableIds(true)
     }
 
-    override fun getItemId(position: Int): Long = position.toLong()
+    override fun getItemId(position: Int): Long {
+        return when (getItem(position)) {
+            is NotesModel -> (getItem(position) as NotesModel).nttUid.toLong()
+            is BookmarkModel -> (getItem(position) as BookmarkModel).bookmarkUid.toLong()
+            is ContentLecture -> (getItem(position) as ContentLecture).lectureUid.toLong()
+            else -> position.toLong()
+        }
+    }
+
+    override fun getItemViewType(position: Int): Int {
+        return position
+    }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
         val inflater = LayoutInflater.from(parent.context)
