--- app/src/main/java/com/codingblocks/cbonlineapp/util/EndlessService.kt
@@ -1,171 +1,0 @@
-package com.codingblocks.cbonlineapp.util
-
-import android.app.Notification
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.app.PendingIntent
-import android.app.Service
-import android.content.Context
-import android.content.Intent
-import android.graphics.Color
-import android.os.Build
-import android.os.IBinder
-import android.os.PowerManager
-import android.util.Log
-import android.widget.Toast
-import androidx.core.app.NotificationCompat
-import com.codingblocks.cbonlineapp.R
-import com.codingblocks.cbonlineapp.admin.AdminActivity
-import com.codingblocks.cbonlineapp.analytics.AppCrashlyticsWrapper.log
-import com.codingblocks.cbonlineapp.util.extensions.isotomillisecond
-import com.codingblocks.onlineapi.CBOnlineLib
-import com.codingblocks.onlineapi.ResultWrapper
-import com.codingblocks.onlineapi.safeApiCall
-import java.util.Locale
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
-
-class EndlessService : Service() {
-
-    private var wakeLock: PowerManager.WakeLock? = null
-    private var isServiceStarted = false
-
-    override fun onBind(intent: Intent): IBinder? {
-        return null
-    }
-
-    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
-        if (intent != null) {
-            when (intent.action) {
-                Actions.START.name -> startService()
-                Actions.STOP.name -> stopService()
-                else -> log("This should never happen. No action in the received intent")
-            }
-        } else {
-            log(
-                "with a null intent. It has been probably restarted by the system."
-            )
-        }
-        // by returning this we make sure the service is restarted if the system kills the service
-        return START_STICKY
-    }
-
-    override fun onCreate() {
-        super.onCreate()
-        val notification = createNotification()
-        startForeground(1, notification)
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        log("The service has been destroyed".toUpperCase(Locale.ROOT))
-        Toast.makeText(this, "Service destroyed", Toast.LENGTH_SHORT).show()
-    }
-
-    private fun startService() {
-        if (isServiceStarted) return
-        log("Starting the foreground service task")
-        Toast.makeText(this, "Service starting its task", Toast.LENGTH_SHORT).show()
-        isServiceStarted = true
-        setServiceState(this, ServiceState.STARTED)
-
-        // we need this lock so our service gets not affected by Doze Mode
-        wakeLock =
-            (getSystemService(Context.POWER_SERVICE) as PowerManager).run {
-                newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "EndlessService::lock").apply {
-                    acquire(10 * 60 * 1000L /*10 minutes*/)
-                }
-            }
-
-        // we're starting a loop in a coroutine
-        GlobalScope.launch(Dispatchers.IO) {
-            while (isServiceStarted) {
-                launch(Dispatchers.IO) {
-                    pingFakeServer()
-                }
-                delay(10 * 60 * 1000)
-            }
-            log("End of the loop for the service")
-        }
-    }
-
-    private fun stopService() {
-        log("Stopping the foreground service")
-        Toast.makeText(this, "Service stopping", Toast.LENGTH_SHORT).show()
-        try {
-            wakeLock?.let {
-                if (it.isHeld) {
-                    it.release()
-                }
-            }
-            stopForeground(true)
-            stopSelf()
-        } catch (e: Exception) {
-            log("Service stopped without being started: ${e.message}")
-        }
-        isServiceStarted = false
-        setServiceState(this, ServiceState.STOPPED)
-    }
-
-    private suspend fun pingFakeServer() {
-        when (val response = safeApiCall(Dispatchers.IO) { CBOnlineLib.onlineV2JsonApi.getLiveDoubts() }) {
-            is ResultWrapper.Success -> with(response.value) {
-                if (isSuccessful)
-                    if (!body()?.get().isNullOrEmpty()) {
-                        body()?.get()?.get(0)?.apply {
-                            Log.i("Notification Worker", "CurrentTime ${System.currentTimeMillis()} AckTime ${createdAt.isotomillisecond()}")
-                            if ((System.currentTimeMillis() - createdAt.isotomillisecond()) < 900000) {
-                                val notification = createNotification()
-                                val mNotificationManager = getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-                                mNotificationManager.notify(1, notification)
-                            }
-                        }
-                    }
-            }
-        }
-    }
-
-    private fun createNotification(): Notification {
-        val notificationChannelId = "DOUBT SERVICE CHANNEL"
-
-        // depending on the Android API that we're dealing with we will have
-        // to use a specific method to create the notification
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val notificationManager = getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-            val channel = NotificationChannel(
-                notificationChannelId,
-                "Doubts Notifications channel",
-                NotificationManager.IMPORTANCE_HIGH
-            ).let {
-                it.description = "Doubts Service channel"
-                it.enableLights(true)
-                it.lightColor = Color.RED
-                it.enableVibration(true)
-                it.vibrationPattern = longArrayOf(100, 200, 300, 400, 500, 400, 300, 200, 400)
-                it
-            }
-            notificationManager.createNotificationChannel(channel)
-        }
-
-        val pendingIntent: PendingIntent = Intent(this, AdminActivity::class.java).let { notificationIntent ->
-            PendingIntent.getActivity(this, 0, notificationIntent, 0)
-        }
-
-        val builder: Notification.Builder = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) Notification.Builder(
-            this,
-            notificationChannelId
-        ) else Notification.Builder(this)
-
-        return builder
-            .setContentTitle("New Doubts to Resolve !!!!!")
-            .setSmallIcon(R.drawable.ic_conversation)
-            .setContentIntent(pendingIntent)
-            .setContentText("Students are waiting for your response.")
-            .setAutoCancel(true)
-            .setDefaults(NotificationCompat.DEFAULT_VIBRATE)
-            .setPriority(Notification.PRIORITY_HIGH) // for under android 26 compatibility
-            .build()
-    }
-}
