--- app/src/main/java/com/codingblocks/cbonlineapp/util/ServiceTrackerf.kt
@@ -1,35 +1,0 @@
-package com.codingblocks.cbonlineapp.util
-
-import android.content.Context
-import android.content.SharedPreferences
-
-enum class ServiceState {
-    STARTED,
-    STOPPED,
-}
-
-enum class Actions {
-    START,
-    STOP
-}
-
-private const val name = "SPYSERVICE_KEY"
-private const val key = "SPYSERVICE_STATE"
-
-fun setServiceState(context: Context, state: ServiceState) {
-    val sharedPrefs = getPreferences(context)
-    sharedPrefs.edit().let {
-        it.putString(key, state.name)
-        it.apply()
-    }
-}
-
-fun getServiceState(context: Context): ServiceState {
-    val sharedPrefs = getPreferences(context)
-    val value = sharedPrefs.getString(key, ServiceState.STOPPED.name)
-    return ServiceState.valueOf(value.toString())
-}
-
-private fun getPreferences(context: Context): SharedPreferences {
-    return context.getSharedPreferences(name, 0)
-}
