+++ app/src/main/java/com/codingblocks/cbonlineapp/workers/DeleteDownloadsWorker.kt
@@ -7,30 +7,32 @@ import androidx.work.WorkerParameters
 import com.codingblocks.cbonlineapp.database.ContentDao
 import com.codingblocks.cbonlineapp.database.models.ContentModel
 import com.codingblocks.cbonlineapp.util.FileUtils
-import com.codingblocks.cbonlineapp.util.MediaUtils
 import org.koin.core.KoinComponent
 import org.koin.core.inject
 import java.io.File
 import java.util.*
 
-class DeleteDownloadsWorker(context: Context, private val workerParameters: WorkerParameters) :
+
+/**
+ * Worker class to Auto-Delete downloaded files after 15 days.
+ */
+class DeleteDownloadsWorker(context: Context, workerParameters: WorkerParameters) :
     CoroutineWorker(context, workerParameters), KoinComponent {
 
     val contentDao: ContentDao by inject()
 
     override suspend fun doWork(): Result {
-        val list:List<ContentModel> = contentDao.getDownloads(true)
+        val list: List<ContentModel> = contentDao.getDownloads(true)
 
-        list.forEach {item->
+        list.forEach { item ->
             val folderFile = File(applicationContext.getExternalFilesDir(Environment.getDataDirectory().absolutePath), "/${item.contentLecture.lectureId}")
             val fileDate = Date(folderFile.lastModified())
             val calendar = Calendar.getInstance()
             calendar.time = Date(fileDate.time)
             calendar.add(Calendar.DATE, 14)
             val expiringDate = calendar.time
-            //Difference is number of days between expiring date and today.
-            val diff = (expiringDate.time - Date().time)/(24 * 60 * 60 * 1000)
-            if (diff<=0){
+            val diff = (expiringDate.time - Date().time) / (24 * 60 * 60 * 1000)
+            if (diff <= 0) {
                 FileUtils.deleteRecursive(folderFile)
                 contentDao.updateContentWithVideoId(item.contentLecture.lectureId, 0)
             }
